+++ b2sdk/_v2/__init__.py
@@ -62,11 +62,14 @@ from b2sdk.utils import trace_call
 
 # data classes
 
+from b2sdk.file_version import DownloadVersion
+from b2sdk.file_version import DownloadVersionFactory
 from b2sdk.file_version import FileIdAndName
 from b2sdk.file_version import FileVersion
 from b2sdk.file_version import FileVersionFactory
 from b2sdk.large_file.part import Part
 from b2sdk.large_file.unfinished_large_file import UnfinishedLargeFile
+from b2sdk.utils.range_ import Range
 
 # file lock
 
@@ -140,7 +143,6 @@ from b2sdk.transfer.inbound.downloader.parallel import ParallelDownloader
 from b2sdk.transfer.inbound.downloader.parallel import PartToDownload
 from b2sdk.transfer.inbound.downloader.parallel import WriterThread
 from b2sdk.transfer.outbound.progress_reporter import PartProgressReporter
-from b2sdk.transfer.inbound.downloader.range import Range
 from b2sdk.transfer.inbound.downloader.simple import SimpleDownloader
 
 # sync